Implement search autocomplete
Uses a new API endpoint (/search/<query>/), which returns a list with
song and artist names that match the given query.

Since we have to call Song.all() often and it accesses the disk a lot,
we use a cache, provided by flask-caching. The return value of the
method stays in the cache for 1 minute, or until a song is saved or
deleted.

The main search view was also improved to search for greeklish queries
in song slugs, and for greek queries in song bodies.